feat(project-token): add buy_tokens event emission logic

+++ contracts/contracts/project-token/src/events.rs
@@ -1,8 +1,7 @@
-// src/events.rs
 use soroban_sdk::{Env, Symbol, Address};
 
-// Event for a token purchase: records the buyer and the amount
-pub fn buy_event(env: &Env, buyer: &Address, amount: i128) {
-    let topics = (Symbol::new(env, "buy_tokens"), buyer.clone());  // event topics
+/// Publishes an event when tokens are bought
+pub fn buy_event(env: &Env, buyer: &Address, amount: u64) {
+    let topics = (Symbol::new(env, "buy_tokens"), buyer.clone());
     env.events().publish(topics, amount);
 }
